refactor(lists): share the event-to-ParsedVideoData mapper
listVideos and pinnedVideoResolution carried byte-identical ~30-line
mappers from event to ParsedVideoData. Extract one parseVideoDataFromEvent
helper and have both resolvers use it.
