+++ src/lib/pinnedVideoResolution.ts
@@ -1,22 +1,10 @@
 // ABOUTME: Resolves pinned-video coordinates into ordered parsed video data
 
 import type { NostrEvent, NostrFilter } from '@nostrify/nostrify';
-import { SHORT_VIDEO_KIND, VIDEO_KINDS, type ParsedVideoData } from '@/types/video';
+import { VIDEO_KINDS, type ParsedVideoData } from '@/types/video';
+import { parseVideoDataFromEvent } from '@/lib/parsedVideoData';
 import { isNewerParsedVideo } from '@/lib/parsedVideoSelection';
 import { parseVideoCoordinate, videoCoordinateKey } from '@/lib/videoCoordinates';
-import {
-  getLoopCount,
-  getOriginalCommentCount,
-  getOriginalLikeCount,
-  getOriginalRepostCount,
-  getOriginalVineTimestamp,
-  getOriginPlatform,
-  getProofModeData,
-  getThumbnailUrl,
-  getVineId,
-  isVineMigrated,
-  parseVideoEvent,
-} from '@/lib/videoParser';
 
 export function buildPinnedVideoFilters(coordinates: string[]): NostrFilter[] {
   const pubkeyGroups = new Map<string, string[]>();
@@ -37,39 +25,6 @@ export function buildPinnedVideoFilters(coordinates: string[]): NostrFilter[] {
   }));
 }
 
-function parsePinnedVideoEvent(event: NostrEvent): ParsedVideoData | null {
-  const vineId = getVineId(event);
-  if (!vineId) return null;
-
-  const videoEvent = parseVideoEvent(event);
-  if (!videoEvent?.videoMetadata?.url) return null;
-
-  return {
-    id: event.id,
-    pubkey: event.pubkey,
-    kind: SHORT_VIDEO_KIND,
-    createdAt: event.created_at,
-    originalVineTimestamp: getOriginalVineTimestamp(event),
-    content: event.content,
-    videoUrl: videoEvent.videoMetadata.url,
-    fallbackVideoUrls: videoEvent.videoMetadata?.fallbackUrls,
-    hlsUrl: videoEvent.videoMetadata?.hlsUrl,
-    thumbnailUrl: getThumbnailUrl(videoEvent),
-    title: videoEvent.title,
-    duration: videoEvent.videoMetadata?.duration,
-    hashtags: videoEvent.hashtags || [],
-    vineId,
-    loopCount: getLoopCount(event),
-    likeCount: getOriginalLikeCount(event),
-    repostCount: getOriginalRepostCount(event),
-    commentCount: getOriginalCommentCount(event),
-    proofMode: getProofModeData(event),
-    origin: getOriginPlatform(event),
-    isVineMigrated: isVineMigrated(event),
-    reposts: [],
-  };
-}
-
 export function resolvePinnedVideosFromEvents(
   coordinates: string[],
   events: NostrEvent[],
@@ -77,7 +32,7 @@ export function resolvePinnedVideosFromEvents(
   const videoMap = new Map<string, ParsedVideoData>();
 
   for (const event of events) {
-    const video = parsePinnedVideoEvent(event);
+    const video = parseVideoDataFromEvent(event);
     if (!video?.vineId) continue;
 
     const key = videoCoordinateKey({ pubkey: event.pubkey, dTag: video.vineId });
